+++ src/test/java/com/graphaware/neo4j/config/MultipleNeo4jVersionsTest.java
@@ -34,7 +34,7 @@ public abstract class MultipleNeo4jVersionsTest {
     protected static String HEAP_SIZE = "256M";
 
     public static Stream<String> neo4jVersions() {
-        return Stream.of("4.4.19", "5.7.0");
+        return Stream.of("4.4.21", "5.9.0");
     }
 
     protected static String heapSizeSetting(Neo4jVersion version) {